+++ src/main/java/ppztw/AdvertBoard/Controller/CategoryController.java
@@ -9,19 +9,19 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import ppztw.AdvertBoard.Exception.ResourceNotFoundException;
-import ppztw.AdvertBoard.Model.Advert;
-import ppztw.AdvertBoard.Model.Category;
-import ppztw.AdvertBoard.Model.CategoryInfo;
-import ppztw.AdvertBoard.Model.InfoType;
+import ppztw.AdvertBoard.Model.Advert.Advert;
+import ppztw.AdvertBoard.Model.Advert.Category;
+import ppztw.AdvertBoard.Model.Advert.CategoryInfo;
+import ppztw.AdvertBoard.Model.Advert.InfoType;
+import ppztw.AdvertBoard.Payload.Advert.CreateCategoryRequest;
 import ppztw.AdvertBoard.Payload.ApiResponse;
-import ppztw.AdvertBoard.Payload.CreateCategoryRequest;
-import ppztw.AdvertBoard.Repository.AdvertRepository;
-import ppztw.AdvertBoard.Repository.CategoryRepository;
+import ppztw.AdvertBoard.Repository.Advert.AdvertRepository;
+import ppztw.AdvertBoard.Repository.Advert.CategoryRepository;
 import ppztw.AdvertBoard.Security.CurrentUser;
 import ppztw.AdvertBoard.Security.UserPrincipal;
 import ppztw.AdvertBoard.Util.PageUtils;
 import ppztw.AdvertBoard.View.Advert.AdvertSummaryView;
-import ppztw.AdvertBoard.View.CategoryView;
+import ppztw.AdvertBoard.View.Advert.CategoryView;
 
 import javax.validation.Valid;
 import java.time.LocalDate;
