+++ src/main/java/ppztw/AdvertBoard/Model/Advert/Image.java
@@ -1,4 +1,4 @@
-package ppztw.AdvertBoard.Model;
+package ppztw.AdvertBoard.Model.Advert;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
