+++ src/main/java/ppztw/AdvertBoard/Model/Advert/InfoType.java
@@ -1,4 +1,4 @@
-package ppztw.AdvertBoard.Model;
+package ppztw.AdvertBoard.Model.Advert;
 
 public enum InfoType {
     money,
