+++ src/main/java/ppztw/AdvertBoard/Repository/Advert/CategoryRepository.java
@@ -1,8 +1,8 @@
-package ppztw.AdvertBoard.Repository;
+package ppztw.AdvertBoard.Repository.Advert;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ppztw.AdvertBoard.Model.Category;
+import ppztw.AdvertBoard.Model.Advert.Category;
 
 import java.util.Optional;
