+++ src/main/java/ppztw/AdvertBoard/Repository/Advert/ImageRepository.java
@@ -1,8 +1,8 @@
-package ppztw.AdvertBoard.Repository;
+package ppztw.AdvertBoard.Repository.Advert;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ppztw.AdvertBoard.Model.Image;
+import ppztw.AdvertBoard.Model.Advert.Image;
 
 @Repository
 public interface ImageRepository extends JpaRepository<Image, Long> {
