+++ src/main/java/ppztw/AdvertBoard/Repository/Advert/TagRepository.java
@@ -1,8 +1,8 @@
-package ppztw.AdvertBoard.Repository;
+package ppztw.AdvertBoard.Repository.Advert;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ppztw.AdvertBoard.Model.Tag;
+import ppztw.AdvertBoard.Model.Advert.Tag;
 
 import java.util.Optional;
